+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- Ignore hidden options when coloring help screen.
+
 ## [4.11.3 (2024-11-11)](https://github.com/kdeldycke/click-extra/compare/v4.11.2...v4.11.3)
 
 - Aligns dependencies.